+++ src/main/java/attendance/controller/AttendanceController.java
@@ -8,13 +8,13 @@ import attendance.domain.AttendanceStatuses;
 import attendance.domain.CrewAttendanceHistory;
 import attendance.domain.CrewName;
 import attendance.domain.DateTimeFormatterWrapper;
-import attendance.dto.AttendanceStatusCount;
-import attendance.dto.ModifyAttendanceDto;
-import attendance.dto.RequestModifyAttendanceDto;
 import attendance.exception.AttendanceArgumentException;
-import attendance.view.AttendanceMethod;
 import attendance.view.ConsoleInputView;
 import attendance.view.OutputView;
+import attendance.view.dto.AttendanceMethod;
+import attendance.view.dto.AttendanceStatusCount;
+import attendance.view.dto.ModifyAttendanceDto;
+import attendance.view.dto.RequestModifyAttendanceDto;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
