--- src/main/java/attendance/dto/AttendanceHistoryDto.java
@@ -1,6 +1,0 @@
-package attendance.dto;
-
-import java.util.List;
-
-public record AttendanceHistoryDto(String nickname, List<String> attendanceHistories, String attendanceDismissStatus) {
-}
