--- src/main/java/attendance/exception/AttendanceStateException.java
@@ -1,8 +1,0 @@
-package attendance.exception;
-
-public class AttendanceStateException extends IllegalStateException {
-
-    public AttendanceStateException(String message) {
-        super("[ERROR] " + message);
-    }
-}
