+++ src/main/java/attendance/view/OutputView.java
@@ -3,8 +3,8 @@ package attendance.view;
 import attendance.domain.AttendanceHistory;
 import attendance.domain.AttendanceStatus;
 import attendance.domain.DateTimeFormatterWrapper;
-import attendance.dto.AttendanceStatusCount;
-import attendance.dto.ModifyAttendanceDto;
+import attendance.view.dto.AttendanceStatusCount;
+import attendance.view.dto.ModifyAttendanceDto;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
