+++ src/main/java/attendance/view/dto/AttendanceStatusCount.java
@@ -1,5 +1,4 @@
-package attendance.dto;
+package attendance.view.dto;
 
 public record AttendanceStatusCount(int absence, int late, int attendance) {
-
 }
