+++ src/main/java/attendance/view/dto/RequestModifyAttendanceDto.java
@@ -1,7 +1,8 @@
-package attendance.dto;
+package attendance.view.dto;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 
 public record RequestModifyAttendanceDto(String nickname, LocalDate date, LocalTime time) {
+    
 }
